look, feel and interaction
created a basic broc editor module,
changed the way the looks change in draggable objects from coded changes
to css with transition
implemented transmission of triggers. It still doesn't work goof enough;
but the solution of running the functions with a boolean of wether to
transmit or not is a good idea to simplify the thing.